+++ code.rkt
@@ -19,6 +19,7 @@
          python-code
          racket-code
          rhombus-code
+         rust-code
          swift-code
          wasm-code
          shell-code
@@ -41,6 +42,7 @@
          pythonblock
          racketblock
          rhombusblock
+         rustblock
          swiftblock
          wasmblock
          shellblock
@@ -63,6 +65,7 @@
          pythonblock0
          racketblock0
          rhombusblock0
+         rustblock0
          swiftblock0
          wasmblock0
          shellblock0
@@ -96,6 +99,7 @@
 (define-syntax-rule (python-code . rest) (lang:python-code . rest))
 (define-syntax-rule (racket-code . rest) (lang:racket-code . rest))
 (define-syntax-rule (rhombus-code . rest) (lang:rhombus-code . rest))
+(define-syntax-rule (rust-code . rest) (lang:rust-code . rest))
 (define-syntax-rule (swift-code . rest) (lang:swift-code . rest))
 (define-syntax-rule (wasm-code . rest) (lang:wasm-code . rest))
 (define-syntax-rule (shell-code . rest) (lang:shell-code . rest))
@@ -118,6 +122,7 @@
 (define-syntax-rule (pythonblock . rest) (lang:pythonblock . rest))
 (define-syntax-rule (racketblock . rest) (lang:racketblock . rest))
 (define-syntax-rule (rhombusblock . rest) (lang:rhombusblock . rest))
+(define-syntax-rule (rustblock . rest) (lang:rustblock . rest))
 (define-syntax-rule (swiftblock . rest) (lang:swiftblock . rest))
 (define-syntax-rule (wasmblock . rest) (lang:wasmblock . rest))
 (define-syntax-rule (shellblock . rest) (lang:shellblock . rest))
@@ -140,6 +145,7 @@
 (define-syntax-rule (pythonblock0 . rest) (lang:pythonblock0 . rest))
 (define-syntax-rule (racketblock0 . rest) (lang:racketblock0 . rest))
 (define-syntax-rule (rhombusblock0 . rest) (lang:rhombusblock0 . rest))
+(define-syntax-rule (rustblock0 . rest) (lang:rustblock0 . rest))
 (define-syntax-rule (swiftblock0 . rest) (lang:swiftblock0 . rest))
 (define-syntax-rule (wasmblock0 . rest) (lang:wasmblock0 . rest))
 (define-syntax-rule (shellblock0 . rest) (lang:shellblock0 . rest))