+++ main.rkt
@@ -18,6 +18,7 @@
          python-code
          racket-code
          rhombus-code
+         rust-code
          swift-code
          wasm-code
          shell-code
@@ -40,6 +41,7 @@
          pythonblock
          racketblock
          rhombusblock
+         rustblock
          swiftblock
          wasmblock
          shellblock
@@ -62,6 +64,7 @@
          pythonblock0
          racketblock0
          rhombusblock0
+         rustblock0
          swiftblock0
          wasmblock0
          shellblock0
@@ -95,6 +98,7 @@
 (define-syntax-rule (python-code . rest) (api:python-code . rest))
 (define-syntax-rule (racket-code . rest) (api:racket-code . rest))
 (define-syntax-rule (rhombus-code . rest) (api:rhombus-code . rest))
+(define-syntax-rule (rust-code . rest) (api:rust-code . rest))
 (define-syntax-rule (swift-code . rest) (api:swift-code . rest))
 (define-syntax-rule (wasm-code . rest) (api:wasm-code . rest))
 (define-syntax-rule (shell-code . rest) (api:shell-code . rest))
@@ -117,6 +121,7 @@
 (define-syntax-rule (pythonblock . rest) (api:pythonblock . rest))
 (define-syntax-rule (racketblock . rest) (api:racketblock . rest))
 (define-syntax-rule (rhombusblock . rest) (api:rhombusblock . rest))
+(define-syntax-rule (rustblock . rest) (api:rustblock . rest))
 (define-syntax-rule (swiftblock . rest) (api:swiftblock . rest))
 (define-syntax-rule (wasmblock . rest) (api:wasmblock . rest))
 (define-syntax-rule (shellblock . rest) (api:shellblock . rest))
@@ -139,6 +144,7 @@
 (define-syntax-rule (pythonblock0 . rest) (api:pythonblock0 . rest))
 (define-syntax-rule (racketblock0 . rest) (api:racketblock0 . rest))
 (define-syntax-rule (rhombusblock0 . rest) (api:rhombusblock0 . rest))
+(define-syntax-rule (rustblock0 . rest) (api:rustblock0 . rest))
 (define-syntax-rule (swiftblock0 . rest) (api:swiftblock0 . rest))
 (define-syntax-rule (wasmblock0 . rest) (api:wasmblock0 . rest))
 (define-syntax-rule (shellblock0 . rest) (api:shellblock0 . rest))